requirements.txt
streamlit>=1.20.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pandas>=2.0.0